+++ frontend/src/pages/cans/list/CanList.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, Outlet } from "react-router-dom";
-import { BreadcrumbItem, BreadcrumbList } from "../../UI/Breadcrumb";
+import { BreadcrumbItem, BreadcrumbList } from "../../../components/Breadcrumb";
 import { getCanList } from "./getCanList";
 
 const CanList = () => {
